+++ demo/backend/components/forms/buttons.py
@@ -1,7 +1,9 @@
 from django import forms
+from django.urls import reverse
 
 from crispy_forms_gds.helper import FormHelper
 from crispy_forms_gds.layout import HTML, Button, Div, Layout
+from crispy_forms_gds.templatetags.crispy_forms_gds import button_link, button_start
 
 
 class ButtonsForm(forms.Form):
@@ -13,6 +15,8 @@ class ButtonsForm(forms.Form):
             Div(Button.secondary("find", "Find address")),
             Div(Button.secondary("win", "Win lottery", disabled=True)),
             Div(Button.warning("delete", "Delete account")),
+            Div(HTML(button_link(reverse("home"), "Link"))),
+            Div(HTML(button_start(reverse("home"), "Start now"))),
         )
 
     def get_button(self):